# Add microbenchmarks and manually unroll expand implementations