--- a/src/java.base/share/classes/java/lang/Long.java
+++ b/src/java.base/share/classes/java/lang/Long.java
@@ -2009,7 +2009,7 @@ public final class Long extends Number
         // Bits to move
         long part5 = maskPrefix & mask;
         // Compress mask
-        mask = (mask ^ part4) | (part4 >>> (1 << 4));
+        mask = (mask ^ part5) | (part5 >>> (1 << 4));
         maskCount = maskCount & ~maskPrefix;
 
         maskPrefix = parallelSuffix(maskCount);