--- a/src/java.base/share/classes/java/lang/Long.java
+++ b/src/java.base/share/classes/java/lang/Long.java
@@ -34,6 +34,7 @@ import java.util.Objects;
 import java.util.Optional;
 
 import jdk.internal.misc.CDS;
+import jdk.internal.vm.annotation.ForceInline;
 import jdk.internal.vm.annotation.IntrinsicCandidate;
 
 import static java.lang.String.COMPACT_STRINGS;
@@ -1934,12 +1935,7 @@ public final class Long extends Number
         for (int j = 0; j < 6; j++) {
             // Parallel prefix
             // Mask prefix identifies bits of the mask that have an odd number of 0's to the right
-            long maskPrefix = maskCount  ^ (maskCount  << 1);
-            maskPrefix = maskPrefix ^ (maskPrefix << 2);
-            maskPrefix = maskPrefix ^ (maskPrefix << 4);
-            maskPrefix = maskPrefix ^ (maskPrefix << 8);
-            maskPrefix = maskPrefix ^ (maskPrefix << 16);
-            maskPrefix = maskPrefix ^ (maskPrefix << 32);
+            long maskPrefix = parallelSuffix(maskCount);
             // Bits to move
             long maskMove = maskPrefix & mask;
             // Compress mask
@@ -1979,60 +1975,61 @@ public final class Long extends Number
         long maskCount = ~mask << 1;
         long maskPrefix = parallelSuffix(maskCount);
         // Bits to move
-        long part1 = maskPrefix & mask;
+        long maskMove1 = maskPrefix & mask;
         // Compress mask
-        mask = (mask ^ part1) | (part1 >>> (1 << 0));
+        mask = (mask ^ maskMove1) | (maskMove1 >>> (1 << 0));
         maskCount = maskCount & ~maskPrefix;
 
         maskPrefix = parallelSuffix(maskCount);
         // Bits to move
-        long part2 = maskPrefix & mask;
+        long maskMove2 = maskPrefix & mask;
         // Compress mask
-        mask = (mask ^ part2) | (part2 >>> (1 << 1));
+        mask = (mask ^ maskMove2) | (maskMove2 >>> (1 << 1));
         maskCount = maskCount & ~maskPrefix;
 
         maskPrefix = parallelSuffix(maskCount);
         // Bits to move
-        long part3 = maskPrefix & mask;
+        long maskMove3 = maskPrefix & mask;
         // Compress mask
-        mask = (mask ^ part3) | (part3 >>> (1 << 2));
+        mask = (mask ^ maskMove3) | (maskMove3 >>> (1 << 2));
         maskCount = maskCount & ~maskPrefix;
 
         maskPrefix = parallelSuffix(maskCount);
         // Bits to move
-        long part4 = maskPrefix & mask;
+        long maskMove4 = maskPrefix & mask;
         // Compress mask
-        mask = (mask ^ part4) | (part4 >>> (1 << 3));
+        mask = (mask ^ maskMove4) | (maskMove4 >>> (1 << 3));
         maskCount = maskCount & ~maskPrefix;
 
         maskPrefix = parallelSuffix(maskCount);
         // Bits to move
-        long part5 = maskPrefix & mask;
+        long maskMove5 = maskPrefix & mask;
         // Compress mask
-        mask = (mask ^ part5) | (part5 >>> (1 << 4));
+        mask = (mask ^ maskMove5) | (maskMove5 >>> (1 << 4));
         maskCount = maskCount & ~maskPrefix;
 
         maskPrefix = parallelSuffix(maskCount);
         // Bits to move
-        long part6 = maskPrefix & mask;
+        long maskMove6 = maskPrefix & mask;
 
         long t = i << (1 << 5);
-        i = (i & ~part6) | (t & part6);
+        i = (i & ~maskMove6) | (t & maskMove6);
         t = i << (1 << 4);
-        i = (i & ~part5) | (t & part5);
+        i = (i & ~maskMove5) | (t & maskMove5);
         t = i << (1 << 3);
-        i = (i & ~part4) | (t & part4);
+        i = (i & ~maskMove4) | (t & maskMove4);
         t = i << (1 << 2);
-        i = (i & ~part3) | (t & part3);
+        i = (i & ~maskMove3) | (t & maskMove3);
         t = i << (1 << 1);
-        i = (i & ~part2) | (t & part2);
+        i = (i & ~maskMove2) | (t & maskMove2);
         t = i << (1 << 0);
-        i = (i & ~part1) | (t & part1);
+        i = (i & ~maskMove1) | (t & maskMove1);
 
         // Clear irrelevant bits
         return i & originalMask;
     }
 
+    @ForceInline
     private static long parallelSuffix(long maskCount) {
         long maskPrefix = maskCount ^ (maskCount << 1);
         maskPrefix = maskPrefix ^ (maskPrefix << 2);