Possible improvements to #8115, including simple microbenchmarks

The improvement comes from manually unrolling the `expand` implementations, avoiding allocating an array in the process. 

Benchmarks results, #8115 baseline:
```
Benchmark                                         (size)  Mode  Cnt      Score      Error   Units
Integers.expand                                      500  avgt   15      2.708 ±    0.057   us/op
Integers.expand:-gc.alloc.rate.norm                  500  avgt   15  20000.368 ±    0.051    B/op
Longs.expand                                         500  avgt   15      4.032 ±    0.061   us/op
Longs.expand:-gc.alloc.rate.norm                     500  avgt   15  32000.569 ±    0.133    B/op
```

This PR:
```
Benchmark                            (size)  Mode  Cnt   Score    Error   Units
Integers.expand                         500  avgt   15   1.229 ±  0.001   us/op
Longs.expand                            500  avgt   15   1.484 ±  0.001   us/op
```

More efficient intrinsics will make optimizing this moot on some platforms, but it seems this unrolling also improve interpreted performance and thus reduce warmup costs, which will help applications perform better until the intrinsic kicks in:

Baseline, `-Xint`:
```
Benchmark                            (size)  Mode  Cnt      Score   Error   Units
Integers.expand                         500  avgt   15    411.433 ± 1.141   us/op
Longs.expand                            500  avgt   15    596.307 ± 4.337   us/op
```

Experiment, `-Xint`:
```
Benchmark                            (size)  Mode  Cnt    Score    Error   Units
Integers.expand                         500  avgt   15  342.897 ±  6.289   us/op
Longs.expand                            500  avgt   15  490.286 ±  3.468   us/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [ ] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Integration blocker
&nbsp;⚠️ Dependency #8115 must be integrated first

### Reviewers
 * [Paul Sandoz](https://openjdk.java.net/census#psandoz) (@PaulSandoz - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/8146/head:pull/8146` \
`$ git checkout pull/8146`

Update a local copy of the PR: \
`$ git checkout pull/8146` \
`$ git pull https://git.openjdk.java.net/jdk pull/8146/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 8146`

View PR using the GUI difftool: \
`$ git pr show -t 8146`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/8146.diff">https://git.openjdk.java.net/jdk/pull/8146.diff</a>

</details>